backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.5
djangorestframework==3.14.0
filelock==3.13.1
identify==2.5.35
idna==3.6
kombu==5.3.5
nodeenv==1.8.0
platformdirs==4.2.0
pre-commit==3.6.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pymemcache==4.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
